Server/requirements.txt
# PM2.5 Ghostbuster Python Dependencies

# Core dependencies
paho-mqtt>=1.6.1
influxdb>=5.3.1
geojson>=3.0.1
pytz>=2023.3
python-dotenv>=1.0.0

# Logging and monitoring
structlog>=23.1.0
colorlog>=6.7.0

# Data processing
pandas>=2.0.3
numpy>=1.24.3

# Web framework (optional for API)
flask>=2.3.2
flask-restx>=1.1.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.1

# Development tools
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.1

# Security
cryptography>=41.0.3